Country and Breads
Added a change to the CountryTrackerContext that allows us to build breads. This makes me think that don't need to have more than one migration or context layer.
We are now able to get country and it's associated breads returned. I will have to remove a controller action and redo testing.